Added: xjc.xjcVersionUnsupportedStrategy support
The plugin will fail gradle builds (by default) when it encounters
an XJC tool version that does not have a supported version in its
MANIFEST.MF Specification-Version attribute.

This setting allows the project administrator to override the
failure to request the plugin try on a best-effort basis to reuse
a strategy of a supported version and continue.

Configuring would use the strategy of the latest known supported
XJC tool version:
 xjc.xjcVersionUnsupportedStrategy='auto-resolve'

This reduces the onus on the plugin author and maintenance around the
releases of a new XJC tool.