MavenUriResolver: Warn on multiple matches, clarify the selected one
Previously it would take the first match and silently ignore the rest.

This might lead to silent inconsistent behaviour (selection
order maybe dependant on the gradle run at the time, i.e. not
consistently recreatable in CI) with no warning or mechanism to
diagnose and resolve.

Attempt to warn the user and clarify which item got selected and by
emitting the list provide a way for the user to resolve or ignore after
reading diagnostic warning.