CatalogsWithCleanIntegrationTest: JVM behaviour workaround
This commit fixes the test case when run on Windows platform.

This is a best effort to be both compatible with as many
JVMs as possible and workaround the JAR locking and
URLConnection useCaches issue on Windows platform.

Worst case scenario you may get the old issue back and need to
manually restart the Gradle daemon when appropriate.  But for
common JVMs OpenJDK/Oracle from 8 onwards this has a high
chance of working around the issue for you.